requirements.txt
appdirs==1.4.4
asgiref==3.8.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
idna==3.6
importlib_metadata==7.1.0
lxml==5.1.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
psycopg2==2.9.9
pyee==11.1.0
PyJWT==2.8.0
pyppeteer==2.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.1
requests==2.31.0
sendgrid==6.11.0
setuptools==69.0.3
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
stripe==9.1.0
timedelta==2020.12.3
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
websockets==10.4
wheel==0.42.0
zipp==3.18.1
